+++ src/configs/escrows.ts
@@ -27,7 +27,9 @@ const escrows = {
   ],
   "3.2": [
     '0x5badb0143f69015c5c86cbd9373474a9c8ab713b',
-    '0x3c74782de03c0402d207fe41307fe50fe9b6b5c7'
+    '0x3c74782de03c0402d207fe41307fe50fe9b6b5c7',
+    '0xbe7b40eb3a9d85d3a76142cb637ab824f0d35ead',
+    '0x5fc1316119a1b7cec52a2984c62764343dca70c9'
   ]
 }